Implementation of the deriv module.
Starting with gsl1.5 the diff module was obsoleted. Instead the deriv module was introduced. The deriv module allows to specify a start step size.
So there is not too much difference for the wrapper. Common functions have been placed into diff_deriv_common.c. The diff and deriv offer the modules for the two implementations.