If PyGSL does not set gsl_error handler to off, the installed error handler will now store the arguments passed by gsl_error_handler to a single static variable in the initmodule. The variable errors.error_safe_state allows retrieving the information or resetting it.
Normally the status of all functions are converted to python exceptions.
During this process the saved state is reset to zero. This conversion
checks if the to be returned gsl errno is identical to the stored one.

This is not thread safe but should not crash the python interpreter for
functions releasing the GIL (e.g UFUNC's).