Wrapper for multifit motivated by requirements of reqularisation.
Starting as separate module multifit2. The different functions are
defined as methods to the gsl_multifit_workspace. The functions used for
regularisation store intermediate results in workspace. Thus their
behaviour is a bit more like methods to workspace ....

Comments and feedback, in particular to the interface, are highly welcome.
This interface is still subject to changes



Signed-off-by: Pierre Schnizer <Pierre.Schnizer@helmholtz-berlin.de>