[TASK] support testing versus functionality not provided by debian packages
gsl shall be useable also if some tools functionality changes.
Here actions are provided that support to build and test gsl
with swig or gsl which are not standard in the ubuntu distribution
for testing.

Furthermore the build of the debian packages is cached to avoid
unnecessary rebuilds.

Currently used for

* swig 4.3: Swig_AppendOutput call signature changed from swig 4.2 to 4.3
* gsl 2.8: bspline interface changed (in an incompatible) manner